test: split project onboarding journeys
